set the function woh allow us to calculate summits and facades for the forme

+++ include/cloud.h
@@ -10,8 +10,8 @@
 
 int get_size_cloud(const char* file_name);
 float** get_cloud_point(const char* file_name, int* size, float fact);
-float** generate_cloud_point(unsigned int uiStacks, unsigned int uiSlices, double fA, double fB, 
-							double fC, double e1, double e2, int* size, int xo, int yo, int zo, double r0, double r1);
+float** generate_cloud_point(unsigned int uiStacks, unsigned int uiSlices, float fA, float fB, 
+							float fC, float e1, float e2, int* size, int xo, int yo, int zo, float r0, float r1);
 void igenvalues(float M[3][3], float R[3][3], float* lambda1, float* lambda2, float* lambda3);
 int calculate_matrix_of_rotation(float M[3][3], float R[3][3], float* lambda1, float* lambda2, float* lambda3);
 int calculate_matrix_of_initial_moments(float** cloud, int size, float tx, float ty, float tz, float matrix[3][3]);
